Restructure shared package
