+++ packages/shared/package.json
@@ -2,7 +2,6 @@
   "name": "@pragma-poker/shared",
   "version": "1.0.0",
   "description": "",
-  "main": "dist/index.js",
   "scripts": {
     "build": "tsc -b"
   },
