+++ packages/shared/tsconfig.json
@@ -3,7 +3,7 @@
     "target": "ES2016",
     "module": "commonjs",
     "outDir": "./dist",
-    "rootDir": "./src",
+    "rootDir": ".",
     "strict": true,
     "esModuleInterop": true,
     "composite": true,
@@ -13,6 +13,5 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
   },
-  "include": ["src/**/*"],
   "exclude": ["node_modules", "**/*.test.ts"]
 }
